+++ src/document/transaction/delete/delete.ts
@@ -24,14 +24,12 @@ export async function stageDelete<
 	const record = await getRecord({ context, ref })
 	assert(record, {
 		detail: `Record not found (tagName=${ref.tagName}, id=${ref.id})`,
-		method: 'delete',
 		key: 'ELEMENT_NOT_FOUND',
 		ref,
 	})
 
 	assert(record.parent, {
 		detail: 'Cannot delete root element',
-		method: 'delete',
 		key: 'PROTECTED_ROOT',
 	})
 
@@ -46,7 +44,6 @@ export async function stageDelete<
 	const parentRecord = await getRecord({ context, ref: parentRef })
 	assert(parentRecord, {
 		detail: `Parent record not found (tagName=${record.parent.tagName}, id=${record.parent.id})`,
-		method: 'delete',
 		key: 'ELEMENT_NOT_FOUND',
 	})